requirements.txt
flask
google-genai
python-dotenv
bandit
